Norman/requirements/prod.txt
pi# Everything needed in production

# Flask
Flask==0.12
MarkupSafe==0.23
Werkzeug==0.11.15
Jinja2==2.9.5
itsdangerous==0.24
click>=5.0

#Database
flask-mongoengine


# Migrations
Flask-Migrate==2.0.3

# Forms
Flask-WTF==0.14.2
WTForms==2.1

# Deployment
gunicorn>=19.1.1

# Assets
Flask-Assets==0.12
cssmin>=0.2.0
jsmin>=2.0.11

# Auth
Flask-Login==0.4.0
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar==0.10.1

#Facebook
pymessenger

#chat
pyyaml

#mail
Flask-Mail

#NLTK
regex
inflection


#GoogleAPIs
python-google-places==1.4.0
googlemaps==2.4.6
